update claimRevision name to claim-revision
