+++ server/claim-revision/schema/claim-revision.schema.ts
@@ -2,6 +2,7 @@ import { Prop, Schema, SchemaFactory } from "@nestjs/mongoose";
 import * as mongoose from "mongoose";
 import { Claim } from "../../claim/schemas/claim.schema";
 import { Source } from "../../source/schemas/source.schema";
+import { ClaimReview } from "../../claim-review/schemas/claim-review.schema";
 
 export type ClaimRevisionDocument = ClaimRevision & mongoose.Document;
 
@@ -19,20 +20,38 @@ export class ClaimRevision {
     @Prop({ required: true })
     date: Date;
 
-    @Prop({ required: true })
+    @Prop({ 
+        type: mongoose.Types.ObjectId,
+        required: true,
+        ref: "Claim",
+    })
     claim: Claim;
 
+    @Prop({
+        required: true,
+        validate: {
+            validator: (v) => {
+                return ["speech", "twitter"].indexOf(v) !== -1;
+            },
+        },
+        message: (tag) => `${tag} is not a valid claim type.`,
+    })
+    type: string;
+
+    // TODO: Let's not use the auto-increment yet
+    // mongodb will create a default _id field and 
+    // we can use it for the first version
     @Prop({ required: true })
     revisionId: number;
 }
 
 const ClaimRevisionSchemaRaw = SchemaFactory.createForClass(ClaimRevision);;
 
-ClaimRevisionSchemaRaw.virtual('claims', {
-  ref: 'Claim',
-  localField: '_id',
-  foreignField: 'claimRevision'
-})
+ClaimRevisionSchemaRaw.virtual('reviews', {
+    ref: 'ClaimReview',
+    localField: '_id',
+    foreignField: 'claim'
+});
 
 ClaimRevisionSchemaRaw.virtual('sources', {
     ref: 'Source',
